requirements.txt
alembic==1.8.1
anyio==3.6.1
apache-airflow==2.4.1
apache-airflow-providers-ftp>=0
apache-airflow-providers-http>=0
apache-airflow-providers-imap>=0
apache-airflow-providers-sqlite>=0
apispec==3.3.2
argcomplete==2.0.0
attrs==22.1.0
Babel==2.10.3
blinker==1.5
cachelib==0.9.0
cattrs==22.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.5
colorlog==4.8.0
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.1
cron-descriptor==1.2.31
croniter==1.3.7
cryptography==38.0.1
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
exceptiongroup==1.0.0rc9
Flask==2.2.2
Flask-AppBuilder==4.1.3
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
graphviz==0.20.1
greenlet==1.1.3.post0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
importlib-metadata>=0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
lazy-object-proxy==1.7.1
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.1
mdurl==0.1.2
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.0.0
prison==0.2.1
psutil==5.9.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-daemon==2.3.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.4
pytzdata==2020.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.0
rfc3986==1.5.0
rich==12.6.0
scikit-learn==1.0.2
scipy>=0
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy>=0
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.3
swagger-ui-bundle==0.0.9
tabulate==0.9.0
tenacity==8.1.0
termcolor==2.0.1
text-unidecode==1.3
threadpoolctl==3.1.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==2.3.3
zipp==3.9.0
psycopg2-binary==2.9.4